// src/models/user-model.ts
export class userModel {
    name: string;
    lastName: string;
    photoUrl: any;
    comercialAdress: string;
    numberComercial: number;
    complementAdress: any;
    cep: any;
    city: string;
    state: string;
    comercialPhone: any;
    celPhone: any;
    email: any;

    v_name: any;
    v_lastName: any; 
    v_comercialAdress:any;
    v_numberComercial: any;
    v_complementAdress: any;
    v_cep: any; 
    v_city: any;
    v_state: any; 
    v_comercialPhone: any;
    v_celPhone: any; 
    v_email: any; 
    v_photoUrl: any;
    valid: boolean;
} 